requirements.txt
streamlit>=1.40.0
langgraph>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.2.0
exa_py>=1.0.0
python-dotenv>=1.0.0

